requirements.txt
beautifulsoup4==4.12.3
line-bot-sdk==3.14.2
openai==1.55.0
pandas==2.2.3
requests==2.32.3
